Checkpoint commit - VM is working again with revised data structures.  Still investigating instruction design.
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/FreeMat5@4385 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libFM/Assembler.hpp
@@ -12,7 +12,7 @@ class Assembler
 {
   std::vector<BasicBlock*> _postorder;
   CodeBlock *_code;
-  std::vector<uint32_t> _vm_codes;
+  std::vector<uint64_t> _vm_codes;
   void depthFirstSearch(BasicBlock* p);
   void computeJumpOffsets();
 public:
